improve naming & logging
