+++ src/resolvers/Query/scanIpsTable.ts
@@ -1,0 +1,32 @@
+import { DocumentClient } from "aws-sdk/clients/dynamodb"
+
+import { ResolverFn } from "resolvers/ResolverFn"
+
+import { IPS } from "../index"
+
+export const scanIpsTable: ResolverFn = async (
+  obj,
+  args,
+  context,
+  { fieldName, parentType }
+) => {
+  const { docClient } = context
+  const table = IPS
+
+  const params: DocumentClient.ScanInput = {
+    TableName: table,
+    Limit: 50,
+  }
+
+  return await docClient
+    .scan(params)
+    .promise()
+    .then(res => {
+      console.info(parentType.name, fieldName, res)
+      return res.Items
+    })
+    .catch(err => {
+      console.error(parentType.name, fieldName, err)
+      throw err
+    })
+}
